+++ Osiris/Hacks/Tickbase.h
@@ -5,9 +5,11 @@ struct UserCmd;
 namespace Tickbase
 {
 	void shiftTicks(int, UserCmd*, bool = false) noexcept;
-	void run(UserCmd*) noexcept;
-	bool canShift(int ticks, bool shiftanyways) noexcept;
+	void run(UserCmd* cmd, bool& sendPacket) noexcept;
+//	bool shouldRecharge(UserCmd* cmd) noexcept;
+	//bool canShift(int ticks, bool shiftanyways) noexcept;
 
+	static int lastShift{ 0 };
 	struct Tick
 	{
 		int	maxUsercmdProcessticks{ 17 }; //on valve servers this is 8 ticks, always do +1 command
