+++ Osiris/Hooks.cpp
@@ -52,9 +52,10 @@
 #include "SDK/Beams.h"
 #include "extraHooks.h"
 #include "Hacks/Animations.h"
-#include "SDK/Tickbase.h"
+#include "Hacks/Tickbase.h"
 #include "Memory.h"
 #include "Interfaces.h"
+#include "Hacks/Debug.h"
 #include "SDK/Input.h"
 #include "SDK/InputSystem.h"
 #include "SDK/StudioRender.h"
@@ -223,7 +224,6 @@ static bool __stdcall createMove(float inputSampleTime, UserCmd* cmd) noexcept
 
     Aimbot::run(cmd);
     Ragebot::run(cmd, rageBestDmg, rageBestChance, wallbangVector);
-    Tickbase::run(cmd, sendPacket);//run this after ragebot
     Triggerbot::run(cmd);
     Backtrack::run(cmd);
     Misc::edgejump(cmd);
@@ -286,7 +286,7 @@ static bool __stdcall createMove(float inputSampleTime, UserCmd* cmd) noexcept
 
     Animations::update(cmd, sendPacket);
     Animations::fake();
-    Tickbase::run(cmd);//run this after ragebot
+	Tickbase::run(cmd, sendPacket);//run this after ragebot
     return false;
 }
 
@@ -746,74 +746,6 @@ static bool __fastcall WriteUsercmdDeltaToBuffer(void* ecx, void* edx, int slot,
 
     return true;
 }
-
-void WriteUsercmd(void* buf, UserCmd* in, UserCmd* out)
-{
-    static DWORD WriteUsercmdF = (DWORD)memory->WriteUsercmd;
-
-    __asm
-    {
-        mov ecx, buf
-        mov edx, in
-        push out
-        call WriteUsercmdF
-        add esp, 4
-    }
-}
-
-
-static bool __fastcall WriteUsercmdDeltaToBuffer(void* ecx, void* edx, int slot, void* buffer, int from, int to, bool isnewcommand) noexcept
-{
-    auto original = hooks->client.getOriginal<bool, 24, int, void*, int, int, bool>(slot, buffer, from, to, isnewcommand);
-
-    if (_ReturnAddress() == memory->WriteUsercmdDeltaToBufferReturn || Tickbase::tick->tickshift <= 0 || !memory->clientState)
-        return original(ecx, slot, buffer, from, to, isnewcommand);
-
-    if (from != -1)
-        return true;
-
-    int* numBackupCommands = (int*)(reinterpret_cast <uintptr_t> (buffer) - 0x30);
-    int* numNewCommands = (int*)(reinterpret_cast <uintptr_t> (buffer) - 0x2C);
-
-    int32_t newcommands = *numNewCommands;
-
-    int nextcommmand = memory->clientState->lastOutgoingCommand + memory->clientState->chokedCommands + 1;
-    int totalcommands = std::min(Tickbase::tick->tickshift, Tickbase::tick->maxUsercmdProcessticks);
-    Tickbase::tick->tickshift = 0;
-
-    from = -1;
-    *numNewCommands = totalcommands;
-    *numBackupCommands = 0;
-
-    for (to = nextcommmand - newcommands + 1; to <= nextcommmand; to++)
-    {
-        if (!(original(ecx, slot, buffer, from, to, true)))
-            return false;
-
-        from = to;
-    }
-
-    UserCmd* lastRealCmd = memory->input->GetUserCmd(slot, from);
-    UserCmd fromcmd;
-
-    if (lastRealCmd)
-        fromcmd = *lastRealCmd;
-
-    UserCmd tocmd = fromcmd;
-    tocmd.tickCount += 200;
-    tocmd.commandNumber++;
-
-    for (int i = newcommands; i <= totalcommands; i++)
-    {
-        WriteUsercmd(buffer, &tocmd, &fromcmd);
-        fromcmd = tocmd;
-        tocmd.commandNumber++;
-        tocmd.tickCount++;
-    }
-
-    return true;
-}
-
 
 void Hooks::install() noexcept
 {
