+++ Osiris/Osiris.vcxproj.filters
@@ -176,7 +176,6 @@
     <ClCompile Include="extraHooks.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-
     <ClCompile Include="Hacks\Resolver.cpp">
       <Filter>Hacks</Filter>
     </ClCompile>
@@ -192,7 +191,6 @@
     <ClCompile Include="Hacks\Tickbase.cpp">
       <Filter>Hacks</Filter>
     </ClCompile>
-
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GUI.h">
@@ -519,7 +517,6 @@
     <ClInclude Include="SDK\Angle.h">
       <Filter>SDK</Filter>
     </ClInclude>
-
     <ClInclude Include="Hacks\Resolver.h">
       <Filter>Hacks</Filter>
     </ClInclude>
@@ -531,7 +528,6 @@
     </ClInclude>
     <ClInclude Include="Hacks\Tickbase.h">
       <Filter>Hacks</Filter>
-
     </ClInclude>
   </ItemGroup>
 </Project>
